Revert "Use double precision for the norm_p value in norm reduce GPU kernel"
This reverts commit dc0c83b77db3f33be1d5d5709d1ef5686e56ceec.

+++ paddle/phi/kernels/funcs/reduce_gpu_kernel.h
@@ -1375,7 +1375,7 @@ void ReduceGpuKernel(const KPDevice& dev_ctx,
                      const phi::DenseTensor& x,
                      phi::DenseTensor* y,
                      const std::vector<int>& origin_reduce_dims,
-                     const double norm_p = 1.0) {
+                     const float norm_p = 1.0f) {
   if (x.numel() == 0) {
     dev_ctx.Alloc<Ty>(y);
     return;
